requirements.txt
requests
lxml
urllib3
pdfminer3k
python-docx
